nop bai

+++ ss9_aop_exception_handling/bai_tap/books/src/main/java/com/example/books/repository/IBookRepository.java
@@ -3,5 +3,5 @@ package com.example.books.repository;
 import com.example.books.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IBookRepository extends JpaRepository<Book,Integer> {
+public interface IBookRepository extends JpaRepository<Book, Long> {
 }
